Automatic library searching when signing SGX enclaves
This is a usability feature.

Introduce two SGX-specific manifest syntax:

- sgx.search_trusted_libs.[id] = [url]

Specifying where to search libraries as dependencies of the trusted
executable and libraries.

- sgx.trusted_libs.[id] = [url]

A library file which should be trusted, and any dependencies of
the library should be included in searching.

sgx.trusted_libs.xxx and sgx.trusted_files.xxx are similar, except
sgx.trusted_libs.xxx will be searched for libraries, whereas
sgx.trusted_files.xxx are treated as regular files.