LibOS/ipc, async helper: keep refcount to helper thread on exit
put_thread(self) to free resources by itself is problematic because
the exiting thread is still using them.
Actually put_thread(self) logic in ipc/async helper may cause SEGV
depending on who is the last to put_thread().
If the thread is the last one to free thread area, lock/unlock after
freeing thread area causes segv due to debug message.

So the solution is to keep reference count of those two helper threads
by a (usually master) thread so that the (master) thread will be the
last thread to release the reference count. So self destruction can be
avoided.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>