RFC: WIP: LibOS/shim_signal: deliver signal when user program running
This patch is still Work-In-Progress.
The purpose of this patch is to get feedback before going further
and trigger discussion.
I believe the patch is enough to show the idea and how to implement
it.

Deliver signal when executing in user program via signal stack
frame, not direct calling handler within LibOS.

With this patches, (emulated) signal is delivered only when
user program is running by setting up signal stack frame.
If LibOS or PAL is running with asynchronously signal, the (async)
signal is queued and the signal stack frame is created right before
system call is return.
This also opens up the capability to support sigaltstack.

The current implementation of invoking registered signal handler
directly within LibOS is problematic.
It causes several issues below and hard to solve them.

- race condition in LibOS because invoking registered signal handler
  directly in host signal handler even when cpu is executing within
  LibOS. If the blocking emulated system call is executing, it's very
  likely. It incurs much complexity.

- nested signal can't be handled well.
  If async signal(e.g. SIGALARM) is delivered and then synchronous
  signal(e.g. SIGSEGV) within signal handler is delieved,
  The current implementation tries to queue SIGSEGV, and loop in
  SIGSEGV.

- sigaltstack can't be implemented well
  Typically language runtime uses sigaltstack. It's desirable to
  support those features well.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>