Makefile: Use symlink instead of cp to populate Runtime dir
Currently cp to Runtime dir is used. Sometimes it is confusing when
binary re-build under Pal or LibOS. So symlink is better. Use symlink
instead of cp to populate Runtime dir.
  - It's easy to understand which file under Runtime dir comes from
  - It's easy to understand which directory to make
  - If cp is used, it's quite easy to test old binary under Runtime.

This patch create ln_sf rule and use it instead of cp. and applies to
other recipe with cp.
In this context, creating symlink of glibc libraries is mess. this
patch also cleans it up to use Makefile instead of embedding the logic
into python script.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>