Pal/Linux: specify loaded address of exec in child process
When fork emulation, the first exec file must be mapped to the
same address of the parent process.
If ASLR is enabled, the address can be different from the parent
without address specified. Then forked process can result in SEGV.
Pass the address for exec file to be mapped from parent process
to the child process. and use that address.
At the same time, file handle isn't initialized fully in
pal_linux_main(), so is initialized.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>