[Pal/Linux] Load forked executable at the same address as parent
During fork emulation, the first exec file must be mapped at the same
address as the parent process.
In previous implementation if ASLR was enabled, the address could be
different from the parent, which could result in a corrupted state.
Pass the address for exec file to be mapped from parent process
to the child process and use that address.
At the same time, fix a problem with file handle not being
fully initialized in pal_linux_main().