[Pal/Linux-SGX] Raise PAL_EVENT_ILLEGAL to LibOS when syscall instruction
This patch is preparation for LibOS to trap-and-emulate syscall
instruction.

This patch makes Pal/Linux-SGX raise PAL_EVENT_ILLEGAL to LibOS when
syscall or unknown instruction is found so that LibOS can handle it.
This patch also cleans up _DkExceptionHandler and add check if it's in
PAL or not. If it's in PAL, it panics.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>