[LibOS] Ignore POLLERR and POLLHUP in `events` in poll()
See #389.

Per the manpage, POLLERR and POLLUP are returned in `revents` and
ignored in `events`.  Indeed, programs should not be setting, and
typically don't set -- `events` with these values.

The effect of this bug is, among others, that a process will never get a
disconnect event from a peer socket.