bugfix: Listenign unix domain socket pipeid likely always 0.
This is issue #390 on the Graphene project.

In the function shim_do_bind, if the socket has type AF_UNIX, its pipe
id is set to:

    data->pipeid = dent->rel_path.hash >> 32;

It seems that fs/shm_fs_hash.c's hash_one() function returns a 32-bit
hash, depsite having a return type of uint64_t (at least, empirically
this seems to be the case -- I haven't worked through the actual
code), after the bit shift, the listening socket has a pipeid of 0.  The
result is that a process can have only one listening unix domain socket.

I changed the statement to simply omit the bitshfit:

    data->pipeid = dent->rel_path.hash;

There is a similar one line change in shim_do_connect when connecting
to a UNIX domain socket and setting the pipeid for the connection.