bugfix: ocall_sock_send and ocall_sock_recv can overflow stack.
Fixes #442.

In Pal/src/host/Linux/SGX/encalve_ocalls.c, the ocall_sock_recv and
ocall_sock_send will copy the respective buffer arguments to the
untrusted host process by adjusting the stack.  In the case of
ocall_sock_recv, this occurs in the line

    ms->ms_buf = ALLOC_IN_USER(buf, count);

and in the case of ocall_sock_send, the line

    ms->ms_buf = COPY_TO_USER(buf, count);

Note that both ALLOC_IN_USER and COPY_TO_USER will eventually call
enclave_framework.c's sgx_ocalloc function, which adjusts the stack.

The problem, of course, is that if if the buffer is too big, the stack
will overflow.

The fix is to apply the same check that ocall_read and ocall_write do;
that is, if the buffer to be allocated is too large to fit on the stack,
call ocall_alloc_untursted to allocate some untrusted heap memory.