[Pal/regression] Disable memory protection and deallocation tests on SGX
SGX1 does not support changing the protection of a page nor (un-)mapping
of a page.

Therefore we can't test memory protection.

Since the allocation/deallocation API is inspired by the mmap/munmap,
mapping over an already allocated area and freeing an unallocated area
are both fine so testing if a free was successful is also not possible
(without significant changes, like exposing the internal allocation
list).

By using SGX2 features it will be possible to implement this in the
future.