[Pal/Linux-SGX] enclave_entry.S: close window where %rsp is invalid
On sgx entry, there is a window where %rsp is not setup to trusted
area yet. If async event(host signal) arrives during that window,
stack is setup wrongly to untrusted area.
This patch closes that window so that stack within enclave is setup
correctly.
An unused label, ".Lexception_hander" was removed as related small
clean up.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>