[LibOS] dedicated syscall stack
For syscall emulation, use dedicated stack for it.
If application stack is small(e.g. go goroutines), stack overflow happens
and crash. Especially when pushing all the GP registers at the beginning
of syscalldb.
In order to avoid it, allocate dedicated stack area for syscall emulation
and switch them back and forth.

This is compile time option with SHIM_SYSCALL_STACK in shim_defs.h.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>