[Pal/Linux-SGX] Check ECALL_ENCLAVE_START parameters
With this change we check all parameters from the urts which we can
verify (This is only true if the other PR we depend on are merged, see
code comments).

We can't verify some parameters like our 'uid' or 'manifest_name'. Here
we have to take the value the urts has passed to us and need to be
careful wherever we use them.

To keep the verification code simple we change the 'arguments' and
'environments' parameters from argv style arrays with pointers to
strings to a continuous NUL-delimited string.

TODO: Needs fix for enclave range check from PR #544!