[Pal/Linux-SGX] Make state checks in enclave_entry more rigid
This change enforces that:

 1. .Lreturn_from_ocall can only be reached if we have prepared the
     stack for it.

 2. only one ecall is called per thread.

The check before didn't cover all cases. For example it was possible to
call ECALL_ENCLAVE_START with an NULL pointer as ecall_args and then
enclave_entry would jump to .Lreturn_from_ocall on the next enclave
entry even though the stack was never prepared for an ocall.