[Pal/Linux-SGX] Don't call EEXIT if handle_ecall returns
During normal operation handle_ecall does not return. But when invalid
parameters are passed it does. In this case it's not guaranteed that the
parameters for EEXIT are already prepared. Ecalls are not designed to
return and calling OCALL_EXIT at this point is also not easy. So just go
into an endless loop. This case should only ever be reached if the
outside behaves maliciously.