[LibOS] Optimize do_exit_group for multi thread race condition
If we try to running a C program in graphene, it may not be able to exit
normally when running do_exit_group. Because have a status like that:
another thread do clone syscall after do_exit_group, the new thread
can't get the KILL signal from do_kill_proc().
So, we loop in try_process_exit() to wait all thread exit.

Signed-off-by: Zhang Chen <chen.zhang@intel.com>