[Scripts/regression] Improved subprocess handling
This fixes two problems with subprocess handling:

 1. Use subprocess.communicate to avoid that the pipe can fill up.

 2. Use a process group to ensure that we also kill the subprocess'
    children in case of a timeout.

Note: This does not cover the case that the child returns successfully
but leaves other child processes running. AFAICS this is not easy since
subprocess.communicate() wait(2)s for the process so the pid is no
longer guaranteed to be valid to kill. So we leave this to the outside
environment (Jenkins or whatever).