[Pal/Linux-SGX] Fix check for untrusted stack
The untrusted stack value is used in two places: It's restored before
EEXIT and it's used to allocate arguments for ocalls on the untrusted
stack.

For EEXIT we don't need to check it since the processor will block
access after leaving enclave mode.

If we allocate something on the untrusted stack we need to check that
the whole region is outside the enclave not just the initial pointer.
Otherwise it might point to non-enclave memory initially but grow into
the enclave region.