[Pal/Linux-SGX] record signal and handle it later if signal is being serviced
In Pal/Linux-SGX, host signal handler setups signal frame and returns.
And actual handler is invoked in the normal execution context, not
signal context. So if async signal is sent fast repeatedly, the
handler is nested even with same async signal and causes stack over
flow.

Introduce flags in sgx tls to track whether signal is being serviced.
When async signal is delievered, check if singal is already being serviced.
If yes, record the signal and service it after the current handler is done.
Although there still remains a small window to grow signal stack, it's much
better than the current code.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>