[Pal] Fix _DkGetAvailableUserAddressRange() returning incorrect memory range
This function was broken on Linux and FreeBSD because it set the top of user
memory to be too low. In case of PIE executables the exec itself and LibOS
code were loaded outside the user memory range.

USER_ADDRESS_RESERVED macro was used in PAL to add a gap after the user memory
and PAL. This gap was, however, not respected by LibOS which caused incorrect
loading as stated above.