[Pal, LibOS] Unify handling of ELF auxiliary vectors in LibOS and PAL
PAL initializes a subset of ELF aux vectors on binary start. LibOS does
the same on execve(). Previously, the two flows were similar but not
unified. This led to subtle memory corruption bugs. This commit unifies
the requirements for the minimum supported number of ELF aux vectors as
well as the minimum additional space to hold their auxiliary data (such
as AT_RANDOM bytes). As a side-effect, it removes nauxv argument.