[LibOS] Fix memory leak of qstr->oflow string during checkpointing
During checkpointing, qstr->oflow was redirected to point to the memory
region in the checkpoint and the old memory region was not freed. We
cannot keep qstr->oflow pointing to the checkpoint memory area
because it will be freed after migration, leaving the parent with a
dangling pointer. Instead, keep parent's qstr->oflow intact but copy its
content inside the checkpoint and rebase accordingly in the child as
part of checkpoint restore.