[LibOS] resolve race between creation/deletion async helper
There is a race between creating async helper thread and the tread
exiting.
When the asyc helper breaks while-loop, install_async_event() thinks
the helper is alive and skip to create_async_helper().
In this case while async events is installed, there is no async
helper thread.
Check the condition for async helper to exit with lock held.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>