[LibOS] Emulate vfork() via fork()
Previous implementation of vfork() was both buggy (worked only in simple
cases) and not compatible with the new IPC implementation. We take a
shortcut for now and emulate vfork() via fork(): this is allowed by
POSIX and almost all Linux apps correctly use vfork().

This commit also adds LibOS regression test for vfork() and disables
vfork02 regression test in ltp (because it incorrectly uses vfork).