[LibOS/shim/test/native] multithread on exit test
This test attempts to trigger race conditions observed on program exit
from an SGX enclave by creating busy threads, and then having the main
thread exit without joining on the child threads.

Since the underlying behaviors are racy, this test may need to be
executed many times to trigger a crash or hang.

Signed-off-by: Alex Merritt <mail@alexmerritt.us>