[LibOS] teach append_signal() how to handle SIG_IGN/SIGCHLD
This patch is follow up for #832, "[LibOS] Fixing LTP test clone03".
When signal is masked ans handler = SIG_IGN, it shouldn't queue
signal to avoid memory leak.
If signal is SIGCHLD, it can be discarded even if it's masked.

Reference:
- Linux implementation: do_notify_parent() @ linux/kernel/signal.c
- Standards:
  https://pubs.opengroup.org/onlinepubs/9699919799/functions/_Exit.html

With glibc-2.19, kill12 is broken due to the its emulation.
It may or may not success depending on which, SIGCHLD from parent or
child nanosleep, is earlier.
If SIGCHLD is earlyer, it success.
If child nanosleep is earlier, it hangs.
With the following changeset, glibc is updated so that the issue
doesn't exit.

Related:
glibc patch
> commit 8c873bf0190740ac1102e13ff7aeb6c08048abfd
>    Remove signal handling for nanosleep (bug 16364)
>
>    Linux 2.6.32 and forward do not show the issue regarding SysV SIGCHLD
>    vs. SIG_IGN for nanosleep which make it feasible to use it for sleep
>    implementation without requiring any hacking to handle the spurious
>    wake up.  The issue is likely being fixed before 2.6 and git
>    history [1] [2].
>
>    This patch simplifies the sleep code to call nanosleep directly by
>    using the posix default version.  It also removes the early cancellation
>    tests for zero argument, since nanosleep will handle cancellation
>    in this case.
>
>    [1] https://lkml.org/lkml/2004/11/25/5
>    [2] https://lkml.org/lkml/2003/11/8/50
>
>    Checked on x86_64, ppc64le, and aarch64.
>
>    [BZ #16364]
>    * sysdeps/unix/sysv/linux/sleep.c: Remove file
>    * sysdeps/posix/sleep.c (__sleep): Simplify cancellation handling.

Signed-off-by: Isaku Yamahata <isaku.yamahata@gmail.com>