[Pal] For LibOS, fix handling of preload libs
Pal now invokes ELF initializers of preloaded libraries in the order
specified (those in the manifest file in `libos.preload`).

Also fix how CALL_ENTRY is invoked; it should only be invoked on the
LibOS library libsysdb.so. CALL_ENTRY executes the ELF entry point,
which does NOT return to the caller.

In the manifest, user is now required to add `loader.libos` to tell Pal
a substring name of the LibOS library file (retaining it in the
preload list). Adding this allows us to also decouple the actual name of
the LibOS library, and enable it to become position-independent in the
preload list.

For example:

loader.preload = file:libsysdb,file:libfoo.so,file:libbar.so
loader.libos = sysdb

ELF initializers will be invoked in a left-to-right order, with the ELF
entry point invoked on libsysdb.so.

Dependencies between dynamic libraries are NOT understood or caught by
this changeset. If one library depends on the state initialized by
another, the user must currently ensure the correct order.

Signed-off-by: Alex Merritt <mail@alexmerritt.us>