[Pal] Delete '\' removal in read_config
If a backslah ('\') was encountered in a quoted ('"') value in a manifest,
it was getting deleted, but since the manifest is mapped read-only it caused
a segfault.
Removed this faulty behavior, now values are passed exactly as seen in
a manifest file, e.g. `loader.env.XXX = "\xY"Z"` passes a environmental
variable `XXX` with a value `"\xY"Z"`