[Pal/Linux-SGX] Implement new manifest syntax sgx.file_check_policy
This syntax specifies the file chck policy, determining the behavior of
authentication to the opening file. By default, trusted and allowed files
are strictly defined by the syntax sgx.trusted_files.foo and
sgx.allowed_files.foo. If the file check policy is allow_all_but_log, all
files other than trusted and allowed files are deemed as allowed with
a warning message. This is a convenient way to debug the file
authentication when porting wild applications to Graphene-SGX.

Signed-off-by: Jia Zhang <zhang.jia@linux.alibaba.com>