[CI] run pylint as part of Jenkins pipelines (18.04)
The pipeline is failed only when pylint whines about the code actually
touched as part of the pull request.

Ubuntu 16.04 pipelines are not enabled, since they are affected by this
bug, since fixed: https://github.com/PyCQA/pylint/issues/1216.

The `run-pylint` script is provided so it can be reused, maybe in
`.git/hooks/pre-push`.