[Pal/Linux-SGX] Move from using pthread functions to raw clone() syscall
Previously, the Linux-SGX PAL (ab)used pthread functions to create new threads
as part of the clone(CLONE_THREAD) emulation. This is not necessary and in fact
contradicts the idea that PAL should be Glibc/pthread-agnostic. This commit
re-writes all pthread functionality to raw clone() syscall and additional
plumbing of alt stack and GS-based untrusted-PAL TCB. Now the threading
functionality of Linux-SGX PAL is very similar to that of Linux PAL.