[Pal/Linux-SGX] Add exitless system calls
This commit adds the ability to invoke system calls in "exitless" fashion
(aka asynchronously). New manifest option `sgx.rpc_thread_num = X`
instructs Graphene to create X outside-enclave RPC threads that spin-wait
for syscalls (and any OCALLs in general) on a shared queue. In-enclave
threads do not exit the enclave but instead enqueue the syscall in the queue
and wait for it to be processed by an RPC thread. Not to waste CPU resources,
enclave threads first spin (in the hope syscall will return immediately) and
then sleep on a futex.

Omitting `sgx.rpc_thread_num` defaults to the old way of executing syscalls:
exiting the enclave, executing the syscall, and re-entering it again.

The exitless feature works with multi-process apps and correctly handles
signals. Two tests of exitless syscalls were added to regression tests.