[Pal/Linux-SGX] Correctly propagate arguments to _DkHandleExternalEvent()
Previously, Case A in the in-enclave exception handler in enclave_entry.S
incorrectly set RDI register to -EINTR (to indicate interrupted syscall)
of the current CPU context, instead of setting RDI register of the interrupted
CPU context (which is accessed via SGX_GPR_RDI). This led to very rare data
races that hanged the interrupted thread/corrupted its memory, because
-EINTR incorrectly propated to SGX_GPR_RSI that holds external event (signal
number) and this led to buffer overflows.