[LibOS] Always occupy the same process on execve()
The execve() syscall starts a new executable in the *same* process. Previously,
Graphene followed this convention *only* for non-SGX PALs. If PAL was Linux-SGX,
Graphene silently terminated the process and created a new one. It is unclear
why there was such a special case, so this commit removes it: now even SGX PAL
starts a new executable in the same process (this is only true for single-
threaded processes). Also, EXECVE_RTLD is redundant and is removed.

This deviation from standard execve() behavior resulted in the host shell
becoming detached from the Graphene-SGX process. In turn, this led to our
Bash example (on Ubuntu 18.04, bash version 4.4.19) "terminating" early
from the point of view of Jenkins, and SGX-18.04 pipeline failed.