[Pal/{Linux, Linux-SGX}] introduce DkThreadPalStack
Introduce a new API, DkThreadPalStack to return a range for sigalt stack
of the current thread.
This is a work around for quirkness of LibOS signal delivery
so that LibOS(and Pal) can be re-entered by (emulated) app signal handler.
Once LibOS signal delivery is fixed so that LibOS isn't re-entered,
DkThreadPalStack should be removed.