[CI] Have list of files that we don't complain about
Jenkins should complain about new files and those already fixed. After
fixing a file completely, the file path should be removed from the list.

This is helpful for pylint, because some errors, mostly from category
"refactor", are attributed to the line with function's signature and are
liable to be missed by the heuristics.