[Pal/Linux-SGX] Unify SGX_OCALL_PREPARED and SGX_STACK
For historical reasons we ended up with having the state info "we have
prepared our stack for exiting for an ocall" in two places. There has
been SGX_OCALL_PREPARED as a simple Boolean. But also SGX_STACK is set
to the pre ocall stack pointer if we prepared to exit for an ocall
(except the special case OCALL_EXIT); otherwise it is set to 0.

Keeping them in sync required some special handling in
.Lsetup_exception_handler.

This can be simplified by just using the pre ocall stack variable to
detect the state. This commit renames SGX_STACK to SGX_PRE_OCALL_STACK
to be more explicit about its semantic.

To keep track of the special case OCALL_EXIT the Boolean
SGX_OCALL_EXIT_CALLED is added.