[Pal/Linux-SGX] Check for duplicated files only at runtime
Previously, Linux-SGX PAL performed a duplicated-files check in
`register_trusted_file()` both during enclave initialization and
at runtime (when creating a new file). This check involves full
traversal of the trusted-files list which is slow in cases of
huge number of trusted files (50K files). Moreover, this check
is performed twice, for correctness in multi-threaded cases.

This commit keeps this check only during runtime and skips it
during initialization (because manifest is assumed to have no
duplicates). Skipping it significantly improves startup time.