[LibOS] Check if timeout argument is given in semtimedop()
Previously, shim_do_semtimedop() assumed that the timeout argument
is always given (not NULL). This led to a segfault in Glibc 2.31
which emulates semop() as semtimedop(timeout=NULL). This commit
introduces a simple fix and does not try to add comprehensive error
checking (this has to be done for the whole semaphore code).