[LibOS] Introduce test for system() and fix bug with Glibc 2.31
Glibc 2.31 implements system() POSIX function call via a combination
of clone(CLONE_VM|CLONE_VFORK|SIGCHLD, <child_stack_addr>, ...) and
execve(). This special case was not supported by Graphene. Moreover,
Glibc 2.31 allocates a child stack and puts some variables on this
child stack. After clone, the child process tries to access these
variables on the child stack. However, previously Graphene always
assumed that the child process returns on the same (parent) stack.
Therefore, restored RSP in the child did not point to the child stack
but to the parent stack, which triggered segfaults in Glibc.

This commit fixes this by temporarily rewiring the current thread (the
one issuing clone(VFORK, ...)) so that it points to the provided child
stack, and then rewiring it back to the original parent stack after
the clone (in the parent process). The child process will resume
execution with the correct child stack.

To test this bugfix, a LibOS test for system() is introduced. Also,
all LibOS tests are forced to start enclaves at address 0x0. This
removes checkpointing issues on mixes of PIE- and non-PIE executables
(in particular, this happens in the system() test).