LibOS: Prevent signal handler from disappearing due to SA_RESETHAND
This patch prevents a signal handler that has the SA_RESETHAND flag set
from disappearing when a signal is appended. We do this by passing an
additional boolean to the __get_sighandler() function that allows us to
control whether the signal handler may be reset or not. It is only
desirable to reset the signal handler if an actual invocation of the
handler function occurs, not when only asking 'where is the handler?'