[LibOS,Pal] Support Glibc logic of line-buffered/fully-buffered STDOUT
Glibc has special handling of STDOUT: it checks whether STDOUT is
TTY (app writes to a terminal) or non-TTY (app writes to a pipe/file)
and makes STDOUT line-buffered or fully-buffered respectively. Glibc
uses a trick with IOCTL(TCGETS) to find out whether it is TTY or
non-TTY. Previously, Graphene always returned -EINVAL on this IOCTL,
so STDOUT in Graphene was always fully-buffered. This inconsistency
led to e.g. close(STDOUT) without printing out buffered strings.
This commit fixes this by propagating info on TTY/non-TTY from the
host.