[tests] Unescape the \n when writing it into the nginx config file
Unescape the '\n' as a newline when writing it into the nginx config file
rather than writing it as '\n'.

Use printf rather than /bin/echo -e since this seems to work in any
case. 'echo -e' on some bash version may print the '-e' while on others
it needs to be there to unescape the '\n'.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>