[Examples] Patch lmbench for distros without rpc/rpc.h and llseek
Patch lmbench for distros where the is no rpc/rpc.h from glibc but
needs tirpc. Also patch llseek to use lseek64. This is now working
on Fedora and Ubuntu-18.04 and it works also when libtircp is not
installed, so we can patch unconditionally.

Add the new library to the sgx trusted libraries in the manifest.