[PAL] Change 'va_list ap' to 'va_list* ap' for all functions
On powerpc64 the function that is called with a 'va_list ap' parameter
must either process the 'ap' parameter itself or must pass it to other
functions as a pointer. So those 2nd level functions need to have a
parameter 'va_list* ap'. This patch now solves this problem by modifying
all those 2nd level functions to take 'va_list* ap' since this should
be the common solution for all architectures.