[LibOS] Fix bug to set exec handle's dentry to executable file
Previously, find_mount_from_uri() assumed that the URI of executable
is an absolute path. This is not true in general case, so this commit
removes this assumption. This bug fix allows to correctly follow
symlink `/proc/self/exe` (previously it returned -EINVAL). LibOS's
`proc` test is updated to verify this.