Remove duplicate ucontext from shim_types-arch.h
Remove the duplicate ucontext definitions from shim_types-arch.h and use
the one from the PAL instead. LibOS now has access to the inline functions
for copying PAL_CONTEXT to ucontext and vice versa and we use them where
possible.

We need to introduce a ucontext.h for Skeleton. It does need ucontext
to be defined for being able to compile shim_signal.c. The easiest way
to achieve this is to rely on Linux's ucontext.h.

SGX can reuse Linux's ucontext.h and sigcontext.h.