[Pal] Add locking to mbedtls
Mbedtls is most of the time not thread-safe, even doing AESCMAC is not.
Added locking everywhere, just to be safe and not to have to debug such
nonsense again. Unfortunately, only spinlocks are portable enough to be
used here (as this code is expected to work both in Pal and outside).