[Pal/Linux-SGX] Add sched_setaffinity/sched_getaffinity support
This patch adds the feature of setting/getting cpu affinity for current thread.
Applications in Graphene+SGX now is able to retrieve correct cpu topoloy via
CPUID w/ leaf 0BH instruction.
Note that the GSGX limits the number of SGX threads,
the CPUID can only retrieve the Apic ID that the calling thread is running on.

Signed-off-by: Gary <gordon.king@intel.com>