[LibOS,Pal] Add "loader.insecure__use_cmdline_argv" to more manifests
As part of Protected Argv feature, Graphene now requires an explicit
manifest option to pass command-line arguments, either
"loader.argv_src_file" or "loader.insecure__use_cmdline_argv". As
part of that change, most manifests were updated to include these
options, but not all. This commit fixes this for the rest of
manifests.