[Pal] don't depend on absolute PATH for tools
On NixOS/Guix tools are not installed into /usr/bin but come from
are composed as symlink trees from immutable packages in the /nix/store:

  i.e. /home/joerg/.nix-profile/bin/bash -> /nix/store/hgplspyw16amnbhlbi7qa2938w8i0rz6-bash-interactive-4.4-p23/bin/bash

Due it's immutable nature they are a perfect fit for graphene-sgx since
libraries store paths won't change unlike their traditional /usr/lib
equalivalents and hence don't break code signing.

One might argue that using for absolute paths makes it possible without a PATH
set, however this is not really true, since make in this case relies on other
tools. Also the whole script relies on bash to be in PATH for it's shebang.